+++ e2e/internal/devnet/cmd/add-client.go
@@ -20,6 +20,7 @@ func (c *AddClientCmd) Command() *cobra.Command {
 	var routeLivenessEnablePassive bool
 	var routeLivenessEnableActive bool
 	var routeLivenessPeerMetrics bool
+	var routeLivenessDebug bool
 
 	cmd := &cobra.Command{
 		Use:   "add-client",
@@ -36,6 +37,7 @@ func (c *AddClientCmd) Command() *cobra.Command {
 				RouteLivenessEnablePassive: routeLivenessEnablePassive,
 				RouteLivenessEnableActive:  routeLivenessEnableActive,
 				RouteLivenessPeerMetrics:   routeLivenessPeerMetrics,
+				RouteLivenessDebug:         routeLivenessDebug,
 			})
 			if err != nil {
 				return fmt.Errorf("failed to add client: %w", err)
@@ -52,6 +54,7 @@ func (c *AddClientCmd) Command() *cobra.Command {
 	cmd.Flags().BoolVar(&routeLivenessEnablePassive, "route-liveness-enable-passive", false, "Enable route liveness in passive mode (experimental)")
 	cmd.Flags().BoolVar(&routeLivenessEnableActive, "route-liveness-enable-active", false, "Enable route liveness in active mode (experimental)")
 	cmd.Flags().BoolVar(&routeLivenessPeerMetrics, "route-liveness-peer-metrics", false, "Enable per peer metrics for route liveness (high cardinality)")
+	cmd.Flags().BoolVar(&routeLivenessDebug, "route-liveness-debug", false, "Enable debug logging for route liveness")
 
 	return cmd
 }