+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
   - Now sends its version to the controller in the grpc GetConfig call
 - Client
   - Switch to 64 byte latency probes instead of 32 bytes
+  - Route liveness admin-down signalling and ignore stale remote-down messages
 
 ## [v0.6.11](https://github.com/malbeclabs/doublezero/compare/client/v0.6.10...client/v0.6.11) – 2025-11-13